Simplify: rename to generate.py, remove unused link_map
- Rename docs_gen.py → generate.py (better reflects it updates README too)
- Remove unused link_map (extracted sections don't contain internal links)
- Script reduced from 95 to 71 lines
